% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/calc_steepness_prior.R
\name{calc_steepness_prior}
\alias{calc_steepness_prior}
\title{Calculate prior for steepness}
\usage{
calc_steepness_prior(LogLike_zp, h_z, Interpolate = TRUE, Degree = 1,
  Ninterp = 203, StockWeight = rep(1, ncol(LogLike_zp)),
  Downweight_outliers = 1, Remove_at_bounds = FALSE, NburnIn = 1e+06,
  Nsamp = 1e+06, Nthin = 100, RunDir = paste0(getwd(), "/"),
  AdmbDir = system.file("executables", package = "SebastesSteepness"))
}
\arguments{
\item{LogLike_zp}{matrix of log-likelihood values for each species \code{p} and likelihood value \code{z}}

\item{h_z}{Steepness value used to calculate each row of \code{LogLike_zp}}
}
\description{
\code{calc_steepness_prior} calculates a prior for steepness given likelihood profiles for individual species
}

