% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SebastesSteepness.R
\docType{package}
\name{SebastesSteepness}
\alias{SebastesSteepness}
\alias{SebastesSteepness-package}
\title{Calculate prior for steepness}
\description{
SebastesSteepness is used to calculate a prior for steepness for West Coast rockfishes
}

